+++ src/platform/src/common/components/PageTopBar/content_less_header.jsx
