+++ src/platform/src/pages/collocation/collocate_success.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import CollocationSuccessImg from '@/icons/Collocation/collocate_success.svg';
-import ContentLessTopBar from '@/components/TopBar/content_less_header';
+import ContentLessTopBar from '@/components/PageTopBar/content_less_header';
 import Button from '@/components/Button';
 import withAuth from '@/core/utils/protectedRoute';
